pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "flask-dance>=7.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "oauthlib>=3.2.2",
    "openai>=1.82.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
]
